requirements.txt
chromadb==0.3.*
google-auth
google-cloud-bigquery==3.*
google-cloud-storage
langchain
matplotlib==3.*
seaborn
numpy==1.*
openai
pandas==2.*
pydantic==1.*
python-dotenv==1.*
regex
streamlit==1.*
streamlit-chat
tiktoken
db-dtypes
tabulate
requests
cryptography